+++ src/assistant/types.ts
@@ -3,3 +3,9 @@ export type FileContentJson = {
   content: string;
   type?: "text" | "binary";
 };
+
+export type FileContent = {
+  name: string;
+  content: string;
+  type: "text" | "binary";
+};
